refactor: TokenProvider, TokenManager 통합
