+++ src/main/java/roomescape/application/AuthService.java
@@ -15,13 +15,10 @@ import roomescape.exception.RoomescapeException;
 
 @Service
 public class AuthService {
-    private final TokenProvider tokenProvider;
     private final TokenManager tokenManager;
     private final MemberRepository memberRepository;
 
-    public AuthService(TokenProvider tokenProvider, TokenManager tokenManager,
-                       MemberRepository memberRepository) {
-        this.tokenProvider = tokenProvider;
+    public AuthService(TokenManager tokenManager, MemberRepository memberRepository) {
         this.tokenManager = tokenManager;
         this.memberRepository = memberRepository;
     }
@@ -33,12 +30,12 @@ public class AuthService {
             throw new RoomescapeException(RoomescapeErrorCode.BAD_REQUEST, "로그인 회원 정보가 일치하지 않습니다.");
         }
         String payload = String.valueOf(member.getId());
-        String accessToken = tokenProvider.createToken(payload);
+        String accessToken = tokenManager.createToken(payload);
         return new TokenResponse(accessToken);
     }
 
     public MemberResponse findMemberByToken(String token) {
-        String payload = tokenProvider.getPayload(token);
+        String payload = tokenManager.getPayload(token);
         long id = Long.parseLong(payload);
         return MemberResponse.from(getMemberBy(id));
     }
