+++ src/main/java/roomescape/application/TokenManager.java
@@ -4,6 +4,10 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 
 public interface TokenManager {
+    String createToken(String payload);
+
+    String getPayload(String token);
+
     String extractToken(Cookie[] cookies);
 
     void setToken(HttpServletResponse response, String accessToken);
