--- src/main/java/roomescape/application/TokenProvider.java
@@ -1,7 +1,0 @@
-package roomescape.application;
-
-public interface TokenProvider {
-    String createToken(String payload);
-
-    String getPayload(String token);
-}
