+++ src/main/java/roomescape/infrastructure/auth/JwtTokenManager.java
@@ -1,9 +1,15 @@
 package roomescape.infrastructure.auth;
 
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
+import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
+import java.util.Date;
 import java.util.Objects;
+import javax.crypto.SecretKey;
 import org.springframework.stereotype.Component;
 import roomescape.application.TokenManager;
 import roomescape.exception.RoomescapeErrorCode;
@@ -13,6 +19,41 @@ import roomescape.exception.RoomescapeException;
 public class JwtTokenManager implements TokenManager {
     private static final String TOKEN_KEY = "token";
     private static final int ONE_MINUTE = 60;
+
+    private final JwtTokenProperties jwtTokenProperties;
+
+    public JwtTokenManager(JwtTokenProperties jwtTokenProperties) {
+        this.jwtTokenProperties = jwtTokenProperties;
+    }
+
+    public String createToken(String payload) {
+        Date now = new Date();
+        Date validity = new Date(now.getTime() + jwtTokenProperties.getExpireMilliseconds());
+
+        String secretKey = jwtTokenProperties.getSecretKey();
+        SecretKey key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
+        return Jwts.builder()
+                .subject(String.valueOf(payload))
+                .expiration(validity)
+                .signWith(key)
+                .compact();
+    }
+
+    public String getPayload(String token) {
+        String secretString = jwtTokenProperties.getSecretKey();
+        SecretKey key = Keys.hmacShaKeyFor(secretString.getBytes(StandardCharsets.UTF_8));
+
+        try {
+            return Jwts.parser()
+                    .verifyWith(key)
+                    .build()
+                    .parseSignedClaims(token)
+                    .getPayload()
+                    .getSubject();
+        } catch (ExpiredJwtException e) {
+            throw new RoomescapeException(RoomescapeErrorCode.TOKEN_EXPIRED);
+        }
+    }
 
     @Override
     public String extractToken(Cookie[] cookies) {
